moved bulk validation methods to ResponsiveUIChunkValidator
