+++ src/main/java/util/validator/ResponsiveUIChunkValidator.java
@@ -5,6 +5,9 @@ import net.itarray.automotion.internal.DriverFacade;
 import org.openqa.selenium.WebElement;
 
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentSkipListMap;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static net.itarray.automotion.Element.asElement;
 import static net.itarray.automotion.Element.asElements;
@@ -215,5 +218,161 @@ public class ResponsiveUIChunkValidator extends ResponsiveUIValidator implements
     public ResponsiveUIChunkValidator insideOf(WebElement containerElement, String readableContainerName) {
         validateInsideOfContainer(asElement(containerElement), readableContainerName, rootElements);
         return this;
+    }
+
+    private void validateElementsAreNotOverlapped(List<Element> elements) {
+        for (int firstIndex = 0; firstIndex < elements.size(); firstIndex++) {
+            Element first = elements.get(firstIndex);
+            for (int secondIndex = firstIndex+1; secondIndex < elements.size(); secondIndex++) {
+                Element second = elements.get(secondIndex);
+                if (first.overlaps(second)) {
+                    errors.add("Elements are overlapped", first);
+                    break;
+                }
+            }
+        }
+    }
+
+    private void validateGridAlignment(List<Element> elements, int columns, int rows) {
+        ConcurrentSkipListMap<Integer, AtomicLong> map = new ConcurrentSkipListMap<>();
+        for (Element element : elements) {
+            Integer y = element.getY();
+
+            map.putIfAbsent(y, new AtomicLong(0));
+            map.get(y).incrementAndGet();
+        }
+
+        int mapSize = map.size();
+        if (rows > 0) {
+            if (mapSize != rows) {
+                errors.add(String.format("Elements in a grid are not aligned properly. Looks like grid has wrong amount of rows. Expected is %d. Actual is %d", rows, mapSize));
+            }
+        }
+
+        if (columns > 0) {
+            int errorLastLine = 0;
+            int rowCount = 1;
+            for (Map.Entry<Integer, AtomicLong> entry : map.entrySet()) {
+                if (rowCount <= mapSize) {
+                    int actualInARow = entry.getValue().intValue();
+                    if (actualInARow != columns) {
+                        errorLastLine++;
+                        if (errorLastLine > 1) {
+                            errors.add(String.format("Elements in a grid are not aligned properly in row #%d. Expected %d elements in a row. Actually it's %d", rowCount, columns, actualInARow));
+                        }
+                    }
+                    rowCount++;
+                }
+            }
+        }
+    }
+
+    private void validateRightOffsetForChunk(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasEqualRightOffsetAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has not the same right offset as element #%d", i + 1, i + 2), elementToCompare);
+            }
+        }
+    }
+
+    private void validateLeftOffsetForChunk(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasEqualLeftOffsetAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has not the same left offset as element #%d", i + 1, i + 2), elementToCompare);
+            }
+        }
+    }
+
+    private void validateTopOffsetForChunk(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasEqualTopOffsetAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has not the same top offset as element #%d", i + 1, i + 2), elementToCompare);
+            }
+        }
+    }
+
+    private void validateBottomOffsetForChunk(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasEqualBottomOffsetAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has not the same bottom offset as element #%d", i + 1, i + 2), elementToCompare);
+            }
+        }
+    }
+
+    private void validateSameWidth(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasSameWidthAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has different width. Element width is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has different width. Element width is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+        }
+    }
+
+    private void validateSameHeight(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasSameHeightAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has different height. Element height is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has different height. Element height is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+        }
+    }
+
+    private void validateSameSize(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (!element.hasSameSizeAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has different size. Element size is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has different size. Element size is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+
+        }
+    }
+
+    private void validateNotSameSize(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (element.hasSameSizeAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has same size. Element size is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has same size. Element size is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+
+        }
+    }
+
+    private void validateNotSameWidth(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (element.hasSameWidthAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has same width. Element width is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has same width. Element width is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+
+        }
+    }
+
+    private void validateNotSameHeight(List<Element> elements) {
+        for (int i = 0; i < elements.size() - 1; i++) {
+            Element element = elements.get(i);
+            Element elementToCompare = elements.get(i + 1);
+            if (element.hasSameHeightAs(elementToCompare)) {
+                errors.add(String.format("Element #%d has same height. Element height is: [%d, %d]", (i + 1), element.getWidth(), element.getHeight()), element);
+                errors.add(String.format("Element #%d has same height. Element height is: [%d, %d]", (i + 2), elementToCompare.getWidth(), elementToCompare.getHeight()), elementToCompare);
+            }
+        }
     }
 }
